FM-34: Move queries into the messages crate.
